Inventory refactoring (+ add unit tests) (#12)
* Inventory refactoring
* bugfixes
* Add unit tests
* Button and inventory fixes, more tests

+++ Game/Events/Inventory/InventoryDiscardEvent.cs
@@ -6,7 +6,7 @@ namespace UAlbion.Game.Events.Inventory
     [Event("inv:discard")]
     public class InventoryDiscardEvent : InventorySlotEvent
     {
-        public InventoryDiscardEvent(InventoryType inventoryType, int id, ItemSlotId slotId) 
+        public InventoryDiscardEvent(InventoryType inventoryType, ushort id, ItemSlotId slotId) 
             : base(inventoryType, id, slotId) { }
     }
 }
