+++ Game/Events/Inventory/InventoryPickupAllEvent.cs
@@ -6,7 +6,7 @@ namespace UAlbion.Game.Events.Inventory
     [Event("inv:pickup_all", "Pickup entire stack of items from a party member's inventory")]
     public class InventoryPickupAllEvent : InventorySlotEvent
     {
-        public InventoryPickupAllEvent(InventoryType sourceType, int sourceId, ItemSlotId slotId)
+        public InventoryPickupAllEvent(InventoryType sourceType, ushort sourceId, ItemSlotId slotId)
             : base(sourceType, sourceId, slotId) { }
     }
 }
