+++ Game/Events/Inventory/InventorySellEvent.cs
@@ -6,7 +6,7 @@ namespace UAlbion.Game.Events.Inventory
     [Event("inv:sell")]
     public class InventorySellEvent : InventorySlotEvent
     {
-        public InventorySellEvent(InventoryType inventoryType, int id, ItemSlotId slotId) 
+        public InventorySellEvent(InventoryType inventoryType, ushort id, ItemSlotId slotId) 
             : base(inventoryType, id, slotId) { }
     }
 }
